Append a text

+++ 0x15-file_io/2-append_text_to_file.c
@@ -13,8 +13,8 @@
 
 int append_text_to_file(const char *filename, char *text_content)
 {
-	int len = 0;
-	int length = 0;
+	size_t len = 0;
+	size_t length = 0;
 	int myfile;
 
 	if (filename == NULL)
